Move dictionary boundary tests to concat.rs
Address reviewer feedback by moving integration tests from separate file
to arrow-select/src/concat.rs as requested.

Tests verify:
- 256 unique values work with u8 keys (boundary case)
- 257 values correctly fail with u8 keys
- 65,536 unique values work with u16 keys (boundary case)
- Overflow returns error instead of panicking
